+++ supplier_api_tester/supplier_api_tester/tests/product_catalog.py
@@ -21,7 +21,7 @@ def test_get_timeslots_products(api_url, api_key, version=1):
     '''Get product catalog with use_timeslots=true query filter'''
 
     url = f'{api_url}/v{version}/products'
-    response = client(url, api_key, method=requests.get, params={'use_timeslot': True})
+    response = client(url, api_key, method=requests.get, params={'use_timeslots': True})
     products = get_products(response)
     _validate_timeslots(products, use_timeslots=True)
     return TestResult()
@@ -32,7 +32,7 @@ def test_get_no_timeslots_products(api_url, api_key, version=1):
     '''Get product catalog with use_timeslots=false query filter'''
 
     url = f'{api_url}/v{version}/products'
-    response = client(url, api_key, method=requests.get, params={'use_timeslot': True})
+    response = client(url, api_key, method=requests.get, params={'use_timeslots': True})
     products = get_products(response)
     _validate_timeslots(products, use_timeslots=False)
     return TestResult()
@@ -40,5 +40,5 @@ def test_get_no_timeslots_products(api_url, api_key, version=1):
 
 def _validate_timeslots(products, use_timeslots):
     for product in products:
-        if product.use_timeslot != use_timeslots:
+        if product.use_timeslots != use_timeslots:
             raise FailedTest(f'Product {product.id} with non matching use_timeslots returned')