+++ supplier_server_mock/supplier_server/constants.py
@@ -1,8 +1,8 @@
 PRODUCTS = [
-    {'id': 'A300-FX', 'name': 'A300-FX', 'use_timeslot': True},
-    {'id': 'A400-FX', 'name': 'A400-FX', 'use_timeslot': True, 'description': 'Test timeslot product'},
-    {'id': 'A500-FX', 'name': 'A500-FX', 'use_timeslot': False, 'description': 'Test non timeslot product'},
-    {'id': 'A600-FX', 'name': 'A600-FX', 'use_timeslot': False},
+    {'id': 'A300-FX', 'name': 'A300-FX', 'use_timeslots': True},
+    {'id': 'A400-FX', 'name': 'A400-FX', 'use_timeslots': True, 'description': 'Test timeslot product'},
+    {'id': 'A500-FX', 'name': 'A500-FX', 'use_timeslots': False, 'description': 'Test non timeslot product'},
+    {'id': 'A600-FX', 'name': 'A600-FX', 'use_timeslots': False},
 ]
 VARIANTS = ('Adult', 'Child')
 MAX_DATE_RANGE = 6  # in months